Extract AppsScreen
